# src/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "LinkIO API",
      "version": "1.0.0",
      "description": "API for digital dentistry",
      "contact": {
        "name": "Caio Mikoszewski",
        "email": "caio.mikoszewski@arrowdigital.group"
      },
      "license": {
        "name": "AllRights-Attribution-NonCommercial-NoDerivatives 4.1",
        "url": "https://github.com/arrow-digital/license"
      }
    },
    "servers": [
      {
        "url": "https://vastly-cheerful-cougar.ngrok-free.app",
        "description": "NGROK tunnel server"
      },
      {
        "url": "http://localhost:8080",
        "description": "Local server"
      }
    ],
    "paths": {
...
      "/api/models/pandaScan": {
        "get": {
          "summary": "Get all pandaScans",
          "tags": [
            "PandaScan"
          ],
          "responses": {
            "200": {
              "description": "List of pandaScans",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PandaScan"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "post": {
          "summary": "Create a new pandaScan",
          "tags": [
            "PandaScan"
          ],
          "requestBody": {
            "description": "PandaScan object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PandaScan"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Created pandaScan",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PandaScan"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/pandaScan/{id}": {
        "get": {
          "summary": "Get a pandaScan by ID",
          "tags": [
            "PandaScan"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the pandaScan to retrieve",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "PandaScan object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PandaScan"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "put": {
          "summary": "Update a pandaScan by ID",
          "tags": [
            "PandaScan"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the pandaScan to update",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "description": "PandaScan object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PandaScan"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Updated pandaScan",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "delete": {
          "summary": "Delete a pandaScan by ID",
          "tags": [
            "PandaScan"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the pandaScan to delete",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Deleted pandaScan",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/service": {
        "get": {
          "summary": "Get all services",
          "tags": [
            "Service"
          ],
          "responses": {
            "200": {
              "description": "List of services",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Service"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "post": {
          "summary": "Create a new service",
          "tags": [
            "Service"
          ],
          "requestBody": {
            "description": "Service object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Created service",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/service/{id}": {
        "get": {
          "summary": "Get a service by ID",
          "tags": [
            "Service"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the service to retrieve",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Service object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "put": {
          "summary": "Update a service by ID",
          "tags": [
            "Service"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the service to update",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "description": "Service object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Updated service",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "delete": {
          "summary": "Delete a service by ID",
          "tags": [
            "Service"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the service to delete",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Deleted service",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/user": {
        "get": {
          "summary": "Get all users",
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "List of users",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "post": {
          "summary": "Create a new user",
          "tags": [
            "User"
          ],
          "requestBody": {
            "description": "User object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Created user",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/user/{id}": {
        "get": {
          "summary": "Get a user by ID",
          "tags": [
            "User"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the user to retrieve",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "put": {
          "summary": "Update a user by ID",
          "tags": [
            "User"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the user to update",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "description": "User object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Updated user",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "delete": {
          "summary": "Delete a user by ID",
          "tags": [
            "User"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the user to delete",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Deleted user",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "PandaScan": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "code": {
                    "description": "must be a number and is required",
                    "type": "number"
                },
                "color": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "doctorName": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "expires": {
                    "description": "must be a date and is required",
                    "format": "date-time",
                    "type": "string"
                },
                "material": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "name": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "notes": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "number": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "patientName": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "referenceOrder": {
                    "type": "string"
                },
                "referenceService": {
                    "type": "string"
                },
                "size": {
                    "description": "must be a number and is required",
                    "type": "number"
                },
                "sourceName": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "targetCreditCode": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "targetName": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "time": {
                    "description": "must be a date and is required",
                    "format": "date-time",
                    "type": "string"
                },
                "type": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "url": {
                    "description": "must be a string and is required",
                    "type": "string"
                }
            },
            "required": [
                "sourceName",
                "targetName",
                "targetCreditCode",
                "name",
                "type",
                "patientName",
                "doctorName",
                "notes",
                "time",
                "expires",
                "size",
                "url",
                "code",
                "material",
                "color",
                "number"
            ],
            "title": "PandaScan"
        },
        "Service": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "clinic": {
                    "description": "must be an objectId from the clinic collection and is optional",
                    "type": "string"
                },
                "description": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "lab": {
                    "description": "must be an objectId from the lab collection and is optional",
                    "type": "string"
                },
                "name": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "type": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "value": {
                    "description": "must be a number and is optional",
                    "type": "number"
                },
                "valueType": {
                    "description": "must be a string and is required",
                    "type": "string"
                }
            },
            "required": [
                "valueType",
                "type"
            ],
            "title": "Service"
        },
        "User": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "clinic": {
                    "description": "must be an objectId from the clinic collection and is optional",
                    "type": "string"
                },
                "email": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "lab": {
                    "description": "must be an objectId from the lab collection and is optional",
                    "type": "string"
                },
                "password": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "permissions": {
                    "description": "must be an array of strings and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "username": {
                    "description": "must be a string and is required",
                    "type": "string"
                }
            },
            "required": [
                "username",
                "password"
            ],
            "title": "User"
        }
      }
    }
}